+++ src/main.rs
@@ -28,6 +28,8 @@ mod serving;
 mod optimizer;
 mod version;
 mod consts;
+mod reg_ffm;
+mod reg_lr;
 //use crate::regressor::RegressorTrait;
